Refactoring JSONview, formfiller and Navbar
